Move usePreimage to react-hooks
