+++ packages/page-democracy/src/Overview/Propose.tsx
@@ -6,9 +6,8 @@ import type { HexString } from '@polkadot/util/types';
 
 import React, { useCallback, useState } from 'react';
 
-import usePreimage from '@polkadot/app-preimages/usePreimage';
 import { Input, InputAddress, InputBalance, Modal, TxButton } from '@polkadot/react-components';
-import { useApi, useCall } from '@polkadot/react-hooks';
+import { useApi, useCall, usePreimage } from '@polkadot/react-hooks';
 import { Available } from '@polkadot/react-query';
 import { isFunction, isHex } from '@polkadot/util';
