+++ packages/react-hooks/src/types.ts
@@ -5,9 +5,13 @@ import type { ApiPromise } from '@polkadot/api';
 import type { SubmittableExtrinsic } from '@polkadot/api/types';
 import type { DeriveAccountFlags, DeriveAccountRegistration } from '@polkadot/api-derive/types';
 import type { DisplayedJudgement } from '@polkadot/react-components/types';
+import type { Bytes } from '@polkadot/types';
 import type { AccountId, Balance, BlockNumber, Call, Exposure, Hash, RewardDestination, SessionIndex, StakingLedger, ValidatorPrefs } from '@polkadot/types/interfaces';
-import type { IExtrinsic } from '@polkadot/types/types';
+import type { PalletPreimageRequestStatus } from '@polkadot/types/lookup';
+import type { IExtrinsic, Registry } from '@polkadot/types/types';
 import type { KeyringJson$Meta } from '@polkadot/ui-keyring/types';
+import type { BN } from '@polkadot/util';
+import type { HexString } from '@polkadot/util/types';
 
 export type CallParam = any;
 
@@ -166,4 +170,30 @@ export type BatchType = 'all' | 'default';
 export interface BatchOptions {
   max?: number;
   type?: BatchType;
+}
+
+export interface PreimageDeposit {
+  amount: BN;
+  who: string;
+}
+
+export interface PreimageStatus {
+  count: number;
+  deposit?: PreimageDeposit;
+  isCompleted: boolean;
+  proposalHash: HexString;
+  proposalLength?: BN;
+  registry: Registry;
+  status: PalletPreimageRequestStatus | null;
+}
+
+export interface PreimageBytes {
+  bytes?: Bytes | null;
+  proposal?: Call | null;
+  proposalError?: string | null;
+  proposalWarning?: string | null;
+}
+
+export interface Preimage extends PreimageBytes, PreimageStatus {
+  // just the interfaces above
 }
